[Affinity] Improving new affinity system
The absence of affinity is considered as the last affinity rule in a worker.
Therefore, jobs with no affinity will be executed by any worker as long as they have executed all the jobs matching their previous affinity rules.
New tests have been added in the test_script file to assert that this new system is working.
Finally, the doc has been rewritten.